feat: Application クラスを Controller クラスに名称変更（MVC風）
