--- Presentations/Cli/Applications/IApplication.cs
@@ -1,7 +1,0 @@
-﻿namespace Cli.Applications
-{
-    public interface IApplication
-    {
-        void Run(string[] args);
-    }
-}
