+++ Presentations/Cli/Controllers/IController.cs
@@ -1,0 +1,7 @@
+﻿namespace Cli.Controllers
+{
+    public interface IController
+    {
+        void Run(string[] args);
+    }
+}
